Disambiguate call for windows platform commit_hash:3f7d1518427ed170b9ac8f70724290b71801cae4

+++ library/cpp/yt/memory/shared_range.h
@@ -58,7 +58,7 @@ public:
 
     //! Constructs an empty TSharedRange from a nullptr expression.
     TSharedRange(std::nullptr_t)
-        : TRange<T>(nullptr, 0UL)
+        : TRange<T>(nullptr, static_cast<size_t>(0))
         , Holder_(nullptr)
     { }
